+++ both/controllers/models/FacilitySchema.jsx
@@ -113,7 +113,7 @@ FacilitySchema = {
     attachments: {
     	type:[Object],
     	label:"Attachments",
-    	input:"attachments"
+        input:DocAttachments.FileExplorer
     },
     address:{
     	label:"Address",