+++ both/controllers/models/collections/collections.js
@@ -1,7 +1,6 @@
 Teams = ORM.Collection("Teams");
 Issues = ORM.Collection("Issues");
 Facilities = ORM.Collection("Facilities");
-Messages = ORM.Collection("Messages");
 Users = ORM.Collection(Meteor.users);
 
 Files = new FS.Collection("File", {